+++ CHANGES.md
@@ -39,8 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Added ASCII art banner to welcome users
-- Added a dedicated 'banner' command to display the ASCII art
+- Added ASCII art banner that displays on each execution of the tool
 
 ## [0.1.6] - 2025-05-02